+++ tests/snapshot_tests/test_snapshots.py
@@ -162,8 +162,16 @@ def test_list_view(snap_compare):
     )
 
 
-def test_textlog_max_lines(snap_compare):
-    assert snap_compare("snapshot_apps/textlog_max_lines.py", press=[*"abcde"])
+def test_richlog_max_lines(snap_compare):
+    assert snap_compare("snapshot_apps/richlog_max_lines.py", press=[*"abcde"])
+
+
+def test_log_write_lines(snap_compare):
+    assert snap_compare("snapshot_apps/log_write_lines.py")
+
+
+def test_log_write(snap_compare):
+    assert snap_compare("snapshot_apps/log_write.py")
 
 
 def test_fr_units(snap_compare):
@@ -421,8 +429,8 @@ def test_table_markup(snap_compare):
     assert snap_compare(SNAPSHOT_APPS_DIR / "table_markup.py")
 
 
-def test_textlog_scroll(snap_compare):
-    assert snap_compare(SNAPSHOT_APPS_DIR / "textlog_scroll.py")
+def test_richlog_scroll(snap_compare):
+    assert snap_compare(SNAPSHOT_APPS_DIR / "richlog_scroll.py")
 
 
 def test_tabs_invalidate(snap_compare):
@@ -559,7 +567,9 @@ def test_tooltips_in_compound_widgets(snap_compare):
 
     assert snap_compare(SNAPSHOT_APPS_DIR / "tooltips.py", run_before=run_before)
 
+
 # --- textual-dev library preview tests ---
+
 
 def test_textual_dev_border_preview(snap_compare):
     assert snap_compare(SNAPSHOT_APPS_DIR / "dev_previews_border.py", press=["enter"])
@@ -580,8 +590,23 @@ def test_textual_dev_keys_preview(snap_compare):
 def test_notifications_example(snap_compare) -> None:
     assert snap_compare(WIDGET_EXAMPLES_DIR / "toast.py")
 
+
 def test_notifications_through_screens(snap_compare) -> None:
     assert snap_compare(SNAPSHOT_APPS_DIR / "notification_through_screens.py")
 
+
 def test_notifications_through_modes(snap_compare) -> None:
     assert snap_compare(SNAPSHOT_APPS_DIR / "notification_through_modes.py")
+
+
+def test_print_capture(snap_compare) -> None:
+    assert snap_compare(SNAPSHOT_APPS_DIR / "capture_print.py")
+
+
+def test_text_log_blank_write(snap_compare) -> None:
+    assert snap_compare(SNAPSHOT_APPS_DIR / "text_log_blank_write.py")
+
+
+def test_nested_fr(snap_compare) -> None:
+    # https://github.com/Textualize/textual/pull/3059
+    assert snap_compare(SNAPSHOT_APPS_DIR / "nested_fr.py")